+++ Tests/ConanPackageTest/source/MachEmuPackageTest.py
@@ -18,15 +18,10 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-set (lib_name Controller)
+import MachineTestDeps
+import os
+import sys
 
-set (${lib_name}_include_files
-	${include_dir}/${lib_name}/I${lib_name}.h
-)
+from mach_emuPy import __version__
+print("mach_emu python module version:", __version__)
 
-SOURCE_GROUP("Include Files" FILES ${${lib_name}_include_files})
-
-add_library(${lib_name} INTERFACE ${${lib_name}_include_files})
-
-target_sources(${lib_name} INTERFACE FILE_SET HEADERS BASE_DIRS ${include_dir} FILES "${include_dir}/${lib_name}/I${lib_name}.h")
-install(TARGETS ${lib_name} FILE_SET HEADERS)