+++ Utils/source/Utils.cpp
@@ -20,20 +20,15 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
+#include <bit>
+#include <libbase64.h>
+#include <md5.h>
+#include <stdexcept>
 #ifdef ENABLE_ZLIB
 #include <zlib.h>
 #endif
 
 #include "Utils/Utils.h"
-#include "Utils/base64.hpp"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-#include "Utils/md5.h"
-#ifdef __cplusplus
-}
-#endif
 
 namespace MachEmu::Utils
 {
@@ -44,7 +39,14 @@ namespace MachEmu::Utils
 			throw std::invalid_argument("Invalid binary to text encoder parameter");
 		}
 
-		std::string binToTxt;
+		auto encode = [](const char* src, size_t srcLen)
+		{
+			// dst string needs to be at least 4/3 times the size of the input 
+			std::string binToTxt(srcLen * 1.5, '\0');
+			base64_encode(src, srcLen, binToTxt.data(), &srcLen, 0);
+			binToTxt.resize(srcLen);
+			return binToTxt;
+		};
 
 		if (compressor != "none")
 		{
@@ -62,7 +64,7 @@ namespace MachEmu::Utils
 					throw std::runtime_error("Failed to compress binary data");
 				}
 
-				binToTxt = base64::encode_into<std::string>(dst.begin(), dst.begin() + len);
+				return encode(std::bit_cast<const char*>(dst.data()), len);
 			}
 			else
 #endif
@@ -72,10 +74,8 @@ namespace MachEmu::Utils
 		}
 		else
 		{
-			binToTxt = base64::encode_into<std::string>(bin, bin + binLen);
+			return encode(std::bit_cast<const char*>(bin), binLen);
 		}
-
-		return binToTxt;
 	}
 
 	// dst needs to be of a size equal to the uncompressed input - this needs to be determined by external means
@@ -86,13 +86,16 @@ namespace MachEmu::Utils
 			throw std::invalid_argument("Invalid binary to text decoder parameter");
 		}
 
-		auto bin = base64::decode_into<std::vector<uint8_t>>(src.begin(), src.end());
+		std::vector<uint8_t> bin(src.length());
+		auto binLen = bin.size();
+		base64_decode(src.data(), src.length(), std::bit_cast<char*>(bin.data()), &binLen, 0);
+		bin.resize(binLen);
 
 		if (decompressor != "none")
 		{
 			std::vector<uint8_t> dst;
 
-#ifdef ENABLE_ZLIB			
+#ifdef ENABLE_ZLIB
 			if (decompressor == "zlib")
 			{
 				dst.resize(dstSize);
@@ -122,12 +125,10 @@ namespace MachEmu::Utils
 
 	std::array<uint8_t, 16> Md5(uint8_t* input, uint32_t len)
 	{
-		MD5Context ctx;
-		md5Init(&ctx);
-		md5Update(&ctx, input, len);
-		md5Finalize(&ctx);
-		auto d = ctx.digest;
-	
-		return std::array{ d[0], d[1], d[2], d[3], d[4], d[5], d[6], d[7], d[8], d[9], d[10], d[11], d[12], d[13], d[14], d[15] };
+		std::array<uint8_t, MD5::HashBytes> hash;
+		MD5 md5;
+		md5.add(input, len);
+		md5.getHash(hash.data());
+		return hash;
 	}
 } // namespace MachEmu::Utils